requirements.txt
Flask==2.2.2
tensorflow==2.10.0
tensorflow-io==0.27.0
waitress==2.1.2
Pillow==9.4.0
